Refactoring code
* extract find_valid_mx_regions to ntlink_utils.py before implementing filtering script
* Remove duplicate code between ntlink_overlap_sequences.py and ntlink_utils.py
