+++ azure-pipelines.yml
@@ -27,7 +27,7 @@ jobs:
   - script: |
       source activate ntlink_CI
       cd bin
-      pylint ntlink_pair.py ntlink_utils.py ntlink_stitch_paths.py
+      pylint ntlink_pair.py ntlink_utils.py ntlink_stitch_paths.py ntlink_overlap_sequences.py
       cd ..
     displayName: Run pylint
   - script: make -C src all
